Move code into new singleton class
